requirements.txt
aiogram==3.7.0
motor==3.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
asyncio-cron==0.0.10
pytz==2024.1